[refactor] POM-011-refacto : TASK_STATUS sortie du reducer pour passer dans un fichier propre sous un dossier constants - initialTasks sortie du reducer pour passer dans un fichier propre sous un dossier data - sortie du handler d'ajout de task et du useRef de l'id de la prochaine task du composant TaskList vers App
