+++ frontend/src/constants/tasks.js
@@ -1,0 +1,5 @@
+export const TASK_STATUS = {
+    TODO: "To Do",
+    IN_PROGRESS: "In Progress",
+    DONE: "Done"
+};
