+++ frontend/src/reducers/tasks.js
@@ -1,14 +1,4 @@
-export const TASK_STATUS = {
-    TODO: "To Do",
-    IN_PROGRESS: "In Progress",
-    DONE: "Done"
-};
-
-export const initialTasks = [
-    { id: 0, title: "My first Task", status: TASK_STATUS.TODO, desc: "Add my today tasks to the list", actualCycle: 0, nbCycle: 1, isDone: false },
-    { id: 1, title: "My second Task", status: TASK_STATUS.TODO, desc: "Add a task to the cycle", actualCycle: 0, nbCycle: 2, isDone: false },
-    { id: 2, title: "My third Task", status: TASK_STATUS.DONE, desc: "Finish my today tasks", actualCycle: 1, nbCycle: 1, isDone: true }
-];
+import { TASK_STATUS } from "../constants/tasks";
 
 export function tasksReducer(state, action) {
     switch (action.type) {
